+++ src/main/java/com/snowflake/kafka/connector/internal/SnowflakeSinkServiceFactory.java
@@ -90,8 +90,14 @@ public class SnowflakeSinkServiceFactory {
       return this;
     }
 
+    /**
+     * Set Max Retries for Cleaner thread for the respective SnowflakeSinkService instance at runtime.
+     *
+     * @param retries maximum cleaner thread retries
+     * @return Builder
+     */
     public SnowflakeSinkServiceBuilder setMaxCleanerRetries(
-            int retries
+            long retries
     ) {
       this.service.setMaxCleanerRetries(retries);
       logInfo("set Cleaner Retries to {} \n", retries);
